Move hsGDeviceRef to pnNucleusInc.
