+++ Sources/Plasma/NucleusLib/inc/CMakeLists.txt
@@ -3,6 +3,7 @@ include_directories("../../NucleusLib")
 include_directories("../../NucleusLib/inc")
 
 set(pnNucleusInc_HEADERS
+    hsGDeviceRef.h
     hsGMatState.h
     hsGMatState.inl
     hsResMgr.h
