+++ Sources/Plasma/PubUtilLib/plGImage/plBitmap.cpp
@@ -57,8 +57,8 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 
 #include "hsResMgr.h"
 #include "hsStream.h"
+#include "hsGDeviceRef.h"
 #include "pnKeyedObject/plKey.h"
-#include "plPipeline/hsGDeviceRef.h"
 
 
 //// Static Members ///////////////////////////////////////////////////////////
