+++ Sources/Plasma/PubUtilLib/plGImage/plDynamicTextMap.cpp
@@ -58,7 +58,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "hsExceptions.h"
 
 #include "hsMatrix44.h"
-#include "plPipeline/hsGDeviceRef.h"
+#include "hsGDeviceRef.h"
 #include "plMessage/plDynamicTextMsg.h"
 #include "pnKeyedObject/plKey.h"
 #include "plProfile.h"
