+++ Sources/Plasma/PubUtilLib/plGImage/plLODMipmap.cpp
@@ -44,11 +44,10 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "plLODMipmap.h"
 
 #include "hsResMgr.h"
+#include "hsGDeviceRef.h"
 #include "pnKeyedObject/plKey.h"
 #include "pnKeyedObject/plUoid.h"
 #include "pnMessage/plRefMsg.h"
-
-#include "plPipeline/hsGDeviceRef.h"
 
 
 plLODMipmap::plLODMipmap()
